airbyte_client::apis
Module connection_api

Enums

CreateConnectionError

struct for typed errors of method create_connection

DeleteConnectionError

struct for typed errors of method delete_connection

GetConnectionError

struct for typed errors of method get_connection

ListAllConnectionsForWorkspaceError

struct for typed errors of method list_all_connections_for_workspace

ListConnectionsByActorDefinitionError

struct for typed errors of method list_connections_by_actor_definition

ListConnectionsForWorkspaceError

struct for typed errors of method list_connections_for_workspace

ResetConnectionError

struct for typed errors of method reset_connection

ResetConnectionStreamError

struct for typed errors of method reset_connection_stream

SearchConnectionsError

struct for typed errors of method search_connections

SyncConnectionError

struct for typed errors of method sync_connection

UpdateConnectionError

struct for typed errors of method update_connection

Functions

create_connection

delete_connection

get_connection

list_all_connections_for_workspace

List connections for workspace, including deleted connections.

list_connections_by_actor_definition

list_connections_for_workspace

List connections for workspace. Does not return deleted connections.

reset_connection

reset_connection_stream

search_connections

sync_connection

update_connection

Apply a patch-style update to a connection. Only fields present on the update request body will be updated. Note that if a catalog is present in the request body, the connection’s entire catalog will be replaced with the catalog from the request. This means that to modify a single stream, the entire new catalog containing the updated stream needs to be sent.