airbyte_client::apis
Module web_backend_api

Enums

GetStateTypeError

struct for typed errors of method get_state_type

WebBackendCheckUpdatesError

struct for typed errors of method web_backend_check_updates

WebBackendCreateConnectionError

struct for typed errors of method web_backend_create_connection

WebBackendGetConnectionError

struct for typed errors of method web_backend_get_connection

WebBackendGetWorkspaceStateError

struct for typed errors of method web_backend_get_workspace_state

WebBackendListConnectionsForWorkspaceError

struct for typed errors of method web_backend_list_connections_for_workspace

WebBackendListGeographiesError

struct for typed errors of method web_backend_list_geographies

WebBackendUpdateConnectionError

struct for typed errors of method web_backend_update_connection

Functions

get_state_type

web_backend_check_updates

web_backend_create_connection

web_backend_get_connection

web_backend_get_workspace_state

web_backend_list_connections_for_workspace

web_backend_list_geographies

Returns all available geographies in which a data sync can run.

web_backend_update_connection

Apply a patch-style update to a connection. Only fields present on the update request body will be updated. Any operations that lack an ID will be created. Then, the newly created operationId will be applied to the connection along with the rest of the operationIds in the request body. Apply a patch-style update to a connection. Only fields present on the update request body will be updated. Note that if a catalog is present in the request body, the connection’s entire catalog will be replaced with the catalog from the request. This means that to modify a single stream, the entire new catalog containing the updated stream needs to be sent.